+++ lib/src/webview_options.dart
@@ -471,8 +471,10 @@ class AndroidInAppWebViewOptions
   String regexToCancelSubFramesLoading;
 
   ///Set to `true` to enable Flutter's new Hybrid Composition. The default value is `false`.
-  ///Hybrid Composition is supported starting with Flutter v1.20.
-  ///**NOTE**: It is recommended to use Hybrid Composition only on Android 10+ for a release app, as it can cause framerate drops on animations in Android 9 and lower.
+  ///Hybrid Composition is supported starting with Flutter v1.20+.
+  ///
+  ///**NOTE**: It is recommended to use Hybrid Composition only on Android 10+ for a release app,
+  ///as it can cause framerate drops on animations in Android 9 and lower (see [Hybrid-Composition#performance](https://github.com/flutter/flutter/wiki/Hybrid-Composition#performance)).
   bool useHybridComposition;
 
   ///Set to `true` to be able to listen at the [WebView.androidShouldInterceptRequest] event. The default value is `false`.